Added HorizontalCrankHub class. Updates for Standardized C and Linux Compat
git-svn-id: https://tpb3dtg.svn.sourceforge.net/svnroot/tpb3dtg@21 7c74f4f4-39d4-4077-a54d-10deffb394f4

+++ People/Person.cpp
@@ -152,7 +152,7 @@ void Person::Draw ()
    Render (manimations[mMood] );
 }
 
-void Person::Draw (int vx, int vy)
+void Person::Draw (float vx, float vy)
 {
 
    Render (manimations[mMood], vx, vx+8 );
